// src/types/partials/AuditEventEntity.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TReference } from '../partials/Reference';
import { TCoding } from '../partials/Coding';
import { TBase64Binary } from '../simpleTypes/Base64Binary';
import { TAuditEventDetail } from '../partials/AuditEventDetail';

export type TAuditEventEntity = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    what?: TReference;
    type?: TCoding;
    role?: TCoding;
    lifecycle?: TCoding;
    securityLabel?: TCoding[];
    name?: String;
    description?: String;
    query?: TBase64Binary;
    detail?: TAuditEventDetail[];
};